backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
ecdsa==0.19.0
fastapi==0.115.5
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
